Extract layer creation from map provider
